+++ lib/src/flutter_sound_player.dart
@@ -412,7 +412,7 @@ class FlutterSoundPlayer {
   void _removePlayerCallback() {
     if (playerController != null) {
       playerController
-        ..add(null)
+        //..add(null)
         ..close();
       playerController = null;
     }